sidechain_block_search

Trait SidechainInfo

pub trait SidechainInfo<Block: BlockT>: Client<Block> {
    type Error: Error;

    // Required methods
    fn get_slot_of_block(
        &self,
        block_number: NumberFor<Block>,
    ) -> Result<ScSlotNumber, Self::Error>;
    fn get_epoch_of_block(
        &self,
        block_number: NumberFor<Block>,
    ) -> Result<ScEpochNumber, Self::Error>;
}

Interface for retrieving information about slot and epoch of Partner Chain blocks

Required Associated Types

type Error: Error

Error type

Required Methods

fn get_slot_of_block( &self, block_number: NumberFor<Block>, ) -> Result<ScSlotNumber, Self::Error>

Finds the Partner Chain slot number for a given block number

fn get_epoch_of_block( &self, block_number: NumberFor<Block>, ) -> Result<ScEpochNumber, Self::Error>

Finds the Partner Chain eopch number for a given block number